// web/src/api/generated/models/chapter-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SectionResponse } from './section-response';

/**
 * 章节信息响应
 */
export interface ChapterResponse {
    /**
     * 章节ID
     */
    'id'?: number;
    /**
     * 课程ID
     */
    'courseId'?: number;
    /**
     * 章节标题
     */
    'title'?: string;
    /**
     * 章节描述
     */
    'description'?: string;
    /**
     * 排序
     */
    'sort'?: number;
    /**
     * 小节列表
     */
    'sections'?: Array<SectionResponse>;
    /**
     * 创建时间
     */
    'createTime'?: string;
    /**
     * 更新时间
     */
    'updateTime'?: string;
}