// web/src/api/generated/models/course-progress-summary-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 课程进度汇总响应
 */
export interface CourseProgressSummaryResponse {
    /**
     * 课程ID
     */
    'courseId'?: number;
    /**
     * 总小节数
     */
    'totalSections'?: number;
    /**
     * 已完成小节数
     */
    'completedSections'?: number;
    /**
     * 课程整体进度(百分比)
     */
    'overallProgress'?: number;
    /**
     * 完成率(百分比)
     */
    'completionRate'?: number;
}